refactor: extract blockquote node
